Refactoring Color
* Many depends on color, and it's ok...
* Shoes3/shoes3#333
