+++ shoes/canvas.c
@@ -8,6 +8,7 @@
 #include "shoes/ruby.h"
 #include "shoes/world.h"
 #include "shoes/native/native.h"
+#include "shoes/types/color.h"
 #include "shoes/http.h"
 
 #define SETUP_SHAPE() \
