+++ shoes/types/types.h
@@ -1,4 +1,5 @@
 #include "shoes/types/button.h"
+#include "shoes/types/color.h"
 #include "shoes/types/edit_box.h"
 #include "shoes/types/edit_line.h"
 #include "shoes/types/effect.h"
@@ -13,6 +14,7 @@
 
 #define SHOES_TYPES_INIT \
 	shoes_button_init(); \
+	shoes_color_init(); \
 	shoes_edit_box_init(); \
 	shoes_edit_line_init(); \
 	shoes_effect_init(); \
